Edit remove review from get to post method by using form in html

+++ edu/views.py
@@ -278,8 +278,11 @@ def review(request, tutor_id):
 def remove_review(request, review_id, tutor_id):
     review = get_object_or_404(Review, id=review_id)
     reviewer = review.reviewer
-    if request.user.is_authenticated and request.user == reviewer.user  :
-        review.delete()
+    if  request.user.is_authenticated and \
+        request.user == reviewer.user and \
+        request.method == "POST" :
+            review.delete()
+            
     return HttpResponseRedirect(f'/review/{tutor_id}')
 
 def about_group(request):
